refactor(orchestration): extract spec-budget and story-body parse gate from ticket-validator
The #4723 Spec word-budget additions measurably dropped
ticket-validator.js maintainability 0.63pt base-to-head — past the
repo's 0.5pt per-file tolerance (run-level clean-code audit
adhoc-4722-4723, the Medium finding). The known remedy is extraction
to a lib module:

- spec-budget.js now owns SPEC_SOFT_WORD_BUDGET + the advisory
  spec-word-budget pass; as an advisory pass it soft-fails an
  unparseable body to no-finding instead of throwing (the hard parse
  gate still rejects it)
- story-body-gate.js now owns parseStoryBodyOrThrow +
  assertStoryBodiesParse (the Story #4541 parse gate), imported back
  by the validator
- measured MI: ticket-validator.js 93.706 -> 94.654 (+0.95, more than
  reversing the audited regression); spec-budget.js 100.7,
  story-body-gate.js 103.6
- dead-exports-production baseline row for SPEC_SOFT_WORD_BUDGET moves
  to its new home (still test-only by design)
